catalog: Refactor parser into separate file, drop tinyxml.
Use one single method to parse the xml data, and use the more
efficient expat parser for this.
