+++ CMakeLists.txt
@@ -828,6 +828,7 @@ set(
   include/canvasMenu.h
   include/catalog_handler.h
   include/catalog_mgr.h
+  include/catalog_parser.h
   include/chart1.h
   include/ChartDataInputStream.h
   include/chartdb.h
@@ -942,6 +943,7 @@ set(
   src/bbox.cpp
   src/catalog_handler.cpp
   src/catalog_mgr.cpp
+  src/catalog_parser.cpp
   src/CanvasConfig.cpp
   src/canvasMenu.cpp
   src/CanvasOptions.cpp
